Change file extension and check format
Checks that all files in BasketEntry folder comply with JS standard style
