+++ frontend/src/js/BasketEntry/BasketContainer/ItemContainer/DeleteLink.jsx
@@ -1,6 +1,8 @@
 import React from 'react'
 
-const DeleteLink = (props) => <button onClick={props.deleteFunction.bind(null, props.index)}>delete</button>
+const DeleteLink = (props) => <button onClick={props.deleteFunction.bind(null, props.index)}>
+                                delete
+                              </button>
 
 DeleteLink.propTypes = {
   deleteFunction: React.PropTypes.func.isRequired,
