+++ frontend/src/js/BasketEntry/BasketContainer/ItemContainer/DeletedItemContainer.jsx
@@ -1,17 +1,23 @@
 import React from 'react'
-import Link from './Link.js'
-import ItemName from './ItemName.js'
+import Link from './Link.jsx'
+import ItemName from './ItemName.jsx'
 
 class DeletedItemContainer extends React.Component {
 
   render () {
     const item = this.props.itemInfo
     return (
-      <tr>
-        <td colSpan={this.props.span}><ItemName itemName={item.itemName} url={item.url} /></td>
-        <td><Link callback={this.props.restoreFunction} index={this.props.index} linkName='Restore' /></td>
-        <td><Link callback={this.props.removeFunction} index={this.props.index} linkName='Remove' /></td>
-      </tr>
+    <tr>
+      <td colSpan={this.props.span}>
+        <ItemName itemName={item.itemName} url={item.url} />
+      </td>
+      <td>
+        <Link callback={this.props.restoreFunction} index={this.props.index} linkName='Restore' />
+      </td>
+      <td>
+        <Link callback={this.props.removeFunction} index={this.props.index} linkName='Remove' />
+      </td>
+    </tr>
     )
   }
 }
