+++ frontend/src/js/BasketEntry/BasketContainer/ItemContainer/ItemName.jsx
@@ -1,6 +1,8 @@
 import React from 'react'
 
-const ItemName = (props) => <a href={props.url}>{props.itemName}</a>
+const ItemName = (props) => <a href={props.url}>
+                              {props.itemName}
+                            </a>
 
 ItemName.propTypes = {
   url: React.PropTypes.string.isRequired,
