+++ frontend/src/js/BasketEntry/BasketContainer/ItemContainer/Price.jsx
@@ -1,6 +1,9 @@
 import React from 'react'
 
-const Price = (props) => <p>{props.currencySymbol}{props.cost}</p>
+const Price = (props) => <p>
+                           {props.currencySymbol}
+                           {props.cost}
+                         </p>
 
 Price.propTypes = {
   cost: React.PropTypes.number.isRequired,
