+++ frontend/src/js/amazon.jsx
@@ -3,7 +3,7 @@ import { render } from 'react-dom'
 import Home from './../modules/home.jsx'
 import Login from './../modules/login.jsx'
 import ReactDom from 'react-dom'
-import BasketEntry from './BasketEntry/'
+import BasketEntry from './BasketEntry/index.jsx'
 import Category from './itemCategory//Category.jsx'
 import SearchBar from './searchbar/searchbar.jsx'
 import ProductPage from './ProductPage/ProductPage.jsx'
